+++ backend/src/main/java/shabalin_zaitsau/hotel_reservation_system/backend/Infrastructure/Services/HotelServices/EntityLayer/HotelDeleteService.java
@@ -3,8 +3,10 @@ package shabalin_zaitsau.hotel_reservation_system.backend.Infrastructure.Service
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import shabalin_zaitsau.hotel_reservation_system.backend.Domain.Entities.Hotel;
 import shabalin_zaitsau.hotel_reservation_system.backend.Infrastructure.Services.HotelServices.EntityLayer.interfaces.IHotelDeleteService;
 import shabalin_zaitsau.hotel_reservation_system.backend.Infrastructure.Storages.HotelRepository;
+import shabalin_zaitsau.hotel_reservation_system.backend.Utils.Authentication.Validation.PermissionValidator;
 
 import java.util.UUID;
 
@@ -15,10 +17,12 @@ public class HotelDeleteService implements IHotelDeleteService {
 
     private final HotelRepository hotelRepository;
     private final HotelReadService hotelReadService;
+    private final PermissionValidator validator;
 
     @Override
     public void removeHotelById(UUID hotelId) {
-        hotelReadService.fetchHotelById(hotelId);
+        Hotel existingHotel = hotelReadService.fetchHotelById(hotelId);
+        validator.validateAdminPermission(existingHotel.getManagedBy().getAdminId());
         hotelRepository.deleteById(hotelId);
     }
 }
